fix(core/ui): resolve remaining compilation errors and detekt issues
- Remove invalid spacedBy imports from GlowButton and InkButton
  (spacedBy is a function on Arrangement, not a standalone import)
- Fix animateColorAsState import in InkSwitch (moved to androidx.compose.animation)
- Fix Float/Long type mismatch in NeonSlider sparkle phase
- Add missing Stroke import to NeonSlider
- Replace non-existent luminance() with manual luminanceValue() in CoverThemeExtractor
- Remove unused FPS_CAP from GradientMeshOrbs (fixes detekt UnusedPrivateProperty)
- Add @Suppress for CyclomaticComplexMethod on extractTheme

+++ core/ui/src/main/java/app/otakureader/core/ui/components/NeonSlider.kt
@@ -18,6 +18,7 @@ import androidx.compose.ui.geometry.CornerRadius
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.*
+import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.unit.dp
@@ -45,7 +46,7 @@ fun NeonSlider(
             Sparkle(
                 offset = sparkleRandom.nextFloat(),
                 size = 2f + sparkleRandom.nextFloat() * 3f,
-                phase = sparkleRandom.nextFloat() * 1000
+                phase = (sparkleRandom.nextFloat() * 1000).toLong()
             )
         }
     }
